services/user-service/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
pika==1.3.2
redis==5.0.3
hiredis==2.3.2
django-redis==5.4.0
django-cors-headers==4.3.1
gunicorn==21.2.0
drf-yasg==1.21.7
celery==5.3.6
django-health-check==3.18.1
pybreaker==1.0.1
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==24.3.0
